Move fixtures to init, move migration test to test_init.py
