+++ tests/components/system_bridge/test_config_flow.py
@@ -1,130 +1,28 @@
 """Test the System Bridge config flow."""
-from collections.abc import Awaitable, Callable
-from dataclasses import asdict
-from ipaddress import ip_address
-from typing import Any
 from unittest.mock import patch
 
-from systembridgeconnector.const import TYPE_DATA_UPDATE
 from systembridgeconnector.exceptions import (
     AuthenticationException,
     ConnectionClosedException,
     ConnectionErrorException,
 )
-from systembridgemodels.const import MODEL_SYSTEM
-from systembridgemodels.modules.system import System
-from systembridgemodels.response import Response
 
 from homeassistant import config_entries, data_entry_flow
-from homeassistant.components import zeroconf
-from homeassistant.components.system_bridge.config_flow import SystemBridgeConfigFlow
 from homeassistant.components.system_bridge.const import DOMAIN
-from homeassistant.const import CONF_API_KEY, CONF_HOST, CONF_PORT, CONF_TOKEN
 from homeassistant.core import HomeAssistant
 
+from . import (
+    FIXTURE_AUTH_INPUT,
+    FIXTURE_DATA_RESPONSE,
+    FIXTURE_USER_INPUT,
+    FIXTURE_UUID,
+    FIXTURE_ZEROCONF,
+    FIXTURE_ZEROCONF_BAD,
+    FIXTURE_ZEROCONF_INPUT,
+    mock_data_listener,
+)
+
 from tests.common import MockConfigEntry
-
-FIXTURE_MAC_ADDRESS = "aa:bb:cc:dd:ee:ff"
-FIXTURE_UUID = "e91bf575-56f3-4c83-8f42-70ac17adcd33"
-
-FIXTURE_AUTH_INPUT = {CONF_TOKEN: "abc-123-def-456-ghi"}
-
-FIXTURE_USER_INPUT = {
-    CONF_TOKEN: "abc-123-def-456-ghi",
-    CONF_HOST: "test-bridge",
-    CONF_PORT: "9170",
-}
-
-FIXTURE_ZEROCONF_INPUT = {
-    CONF_TOKEN: "abc-123-def-456-ghi",
-    CONF_HOST: "1.1.1.1",
-    CONF_PORT: "9170",
-}
-
-FIXTURE_ZEROCONF = zeroconf.ZeroconfServiceInfo(
-    ip_address=ip_address("1.1.1.1"),
-    ip_addresses=[ip_address("1.1.1.1")],
-    port=9170,
-    hostname="test-bridge.local.",
-    type="_system-bridge._tcp.local.",
-    name="System Bridge - test-bridge._system-bridge._tcp.local.",
-    properties={
-        "address": "http://test-bridge:9170",
-        "fqdn": "test-bridge",
-        "host": "test-bridge",
-        "ip": "1.1.1.1",
-        "mac": FIXTURE_MAC_ADDRESS,
-        "port": "9170",
-        "uuid": FIXTURE_UUID,
-    },
-)
-
-FIXTURE_ZEROCONF_BAD = zeroconf.ZeroconfServiceInfo(
-    ip_address=ip_address("1.1.1.1"),
-    ip_addresses=[ip_address("1.1.1.1")],
-    port=9170,
-    hostname="test-bridge.local.",
-    type="_system-bridge._tcp.local.",
-    name="System Bridge - test-bridge._system-bridge._tcp.local.",
-    properties={
-        "something": "bad",
-    },
-)
-
-
-FIXTURE_SYSTEM = System(
-    boot_time=1,
-    fqdn="",
-    hostname="1.1.1.1",
-    ip_address_4="1.1.1.1",
-    mac_address=FIXTURE_MAC_ADDRESS,
-    platform="",
-    platform_version="",
-    uptime=1,
-    uuid=FIXTURE_UUID,
-    version="",
-    version_latest="",
-    version_newer_available=False,
-    users=[],
-)
-
-FIXTURE_DATA_RESPONSE = Response(
-    id="1234",
-    type=TYPE_DATA_UPDATE,
-    subtype=None,
-    message="Data received",
-    module=MODEL_SYSTEM,
-    data=asdict(FIXTURE_SYSTEM),
-)
-
-FIXTURE_DATA_RESPONSE_BAD = Response(
-    id="1234",
-    type=TYPE_DATA_UPDATE,
-    subtype=None,
-    message="Data received",
-    module=MODEL_SYSTEM,
-    data={},
-)
-
-FIXTURE_DATA_RESPONSE_BAD = Response(
-    id="1234",
-    type=TYPE_DATA_UPDATE,
-    subtype=None,
-    message="Data received",
-    module=MODEL_SYSTEM,
-    data={},
-)
-
-
-async def mock_data_listener(
-    self,
-    callback: Callable[[str, Any], Awaitable[None]] | None = None,
-    _: bool = False,
-):
-    """Mock websocket data listener."""
-    if callback is not None:
-        # Simulate data received from the websocket
-        await callback(MODEL_SYSTEM, FIXTURE_SYSTEM)
 
 
 async def test_show_user_form(hass: HomeAssistant) -> None:
@@ -536,27 +434,3 @@ async def test_zeroconf_bad_zeroconf_info(hass: HomeAssistant) -> None:
 
     assert result["type"] == data_entry_flow.FlowResultType.ABORT
     assert result["reason"] == "unknown"
-
-
-async def test_migration(hass: HomeAssistant) -> None:
-    """Test migration from system_bridge to system_bridge."""
-    config_entry = MockConfigEntry(
-        domain=DOMAIN,
-        unique_id=FIXTURE_UUID,
-        data={
-            CONF_API_KEY: FIXTURE_USER_INPUT[CONF_TOKEN],
-            CONF_HOST: FIXTURE_USER_INPUT[CONF_HOST],
-            CONF_PORT: FIXTURE_USER_INPUT[CONF_PORT],
-        },
-        version=1,
-        minor_version=1,
-    )
-    config_entry.add_to_hass(hass)
-    await hass.config_entries.async_setup(config_entry.entry_id)
-    await hass.async_block_till_done()
-
-    # Check that the version has been updated and the api_key has been moved to token
-    assert config_entry.version == SystemBridgeConfigFlow.VERSION
-    assert config_entry.minor_version == SystemBridgeConfigFlow.MINOR_VERSION
-    assert config_entry.data[CONF_TOKEN] == FIXTURE_USER_INPUT[CONF_TOKEN]
-    assert config_entry.data == FIXTURE_USER_INPUT
